renamed next-setup file
