+++ packages/sitecore-jss-nextjs/src/tests/next-setup.js
